BAU: Move last three digits helper to shared
This will be useful for the reset password request handler, since it will start returning some of the same fields as the CheckUserExists handler. Therefore this is a preparatory commit that moves this functionality to a shared place.
